+++ satella/coding/sequences/iterators.py
@@ -96,7 +96,8 @@ def infinite_counter(start_at: int = 0, step: int = 1) -> tp.Iterator[int]:
     :param start_at: value at which to start counting. It will be yielded as first
     :param step: step by which to progress the counter
     """
-    warnings.warn('This is deprecated, use itertools.count() instead', DeprecationWarning)
+    warnings.warn('This is deprecated and will be removed in Satella 3.0, '
+                  'use itertools.count() instead', DeprecationWarning)
     i = start_at
     while True:
         yield i