+++ satella/coding/concurrent/__init__.py
@@ -3,8 +3,8 @@ from .callablegroup import CallableGroup, CallNoOftenThan
 from .locked_dataset import LockedDataset
 from .locked_structure import LockedStructure
 from .monitor import MonitorList, Monitor, MonitorDict, RMonitor
-from .thread import TerminableThread, Condition
+from .thread import TerminableThread, Condition, SingleStartThread
 
 __all__ = ['LockedDataset', 'Monitor', 'RMonitor', 'CallableGroup', 'TerminableThread',
            'MonitorDict', 'MonitorList', 'Condition', 'LockedStructure', 'AtomicNumber',
-           'CallNoOftenThan']
+           'CallNoOftenThan', 'SingleStartThread']