+++ tests/test_coding/test_concurrent.py
@@ -18,22 +18,22 @@ class TestConcurrent(unittest.TestCase):
 
         def process_atomic_number():
             with Monitor.acquire(a):
-                time.sleep(5)
+                time.sleep(3)
 
         threading.Thread(target=process_atomic_number).start()
 
-        self.assertRaises(WouldWaitMore, lambda: a.wait_until_equal(3, timeout=2))
+        self.assertRaises(WouldWaitMore, lambda: a.wait_until_equal(3, timeout=1))
 
     def test_atomic_number_wait_equal(self):
         an = AtomicNumber()
 
         def process_atomic_number(a):
-            time.sleep(3)
+            time.sleep(1)
             a += 1
 
         threading.Thread(target=process_atomic_number, args=(an,)).start()
-        self.assertRaises(WouldWaitMore, lambda: an.wait_until_equal(2, timeout=2))
-        an.wait_until_equal(1, timeout=2)
+        self.assertRaises(WouldWaitMore, lambda: an.wait_until_equal(2, timeout=0.5))
+        an.wait_until_equal(1, timeout=1)
         an.wait_until_equal(1)
 
         self.assertEqual(str(an), 'AtomicNumber(1)')
@@ -46,6 +46,12 @@ class TestConcurrent(unittest.TestCase):
         self.assertEqual(a, 3)
 
         self.assertIsInstance(a + 2, int)
+        self.assertEqual(a | 4, 7)
+        self.assertEqual(a & 2, 2)
+        self.assertEqual(a ^ 3, 0)
+        self.assertEqual(4 | a, 7)
+        self.assertEqual(2 & a, 2)
+        self.assertEqual(3 ^ a, 0)
         self.assertTrue(a > 1)
         self.assertTrue(a < 4)
         self.assertTrue(a >= 1)
@@ -55,7 +61,17 @@ class TestConcurrent(unittest.TestCase):
         self.assertEqual(a*1, 3)
         self.assertEqual(a/1, 3.0)
         self.assertEqual(a//1, 3)
-        a *= 1
+        self.assertEqual(1+a, 4)
+        self.assertEqual(-1+a, 2)
+        self.assertEqual(1*a, 3)
+        self.assertEqual(1/(a+1), 0.25)
+        self.assertEqual(1//a, 0)
+        self.assertEqual(2**a, 8)
+        a |= 4
+        self.assertEqual(a, 7)
+        a &= 3
+        self.assertEqual(a, 3)
+        a ^= 0
         self.assertEqual(a, 3)
         a /= 1
         self.assertEqual(a, 3.0)
@@ -64,6 +80,7 @@ class TestConcurrent(unittest.TestCase):
         self.assertEqual(a**2, 9)
         a **= 2
         self.assertEqual(a, 9)
+        self.assertEqual(+a, 9)
         self.assertTrue(bool(a))
         self.assertEqual(int(a), 9)
         self.assertEqual(float(a), 9.0)