+++ CHANGELOG.md
@@ -1,2 +1,3 @@
 # v2.9.4
 
+* changed the allowed characters in metric names