+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # v2.9.4
 
 * changed the allowed characters in metric names
+* added a clear deprecation policy