+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 * changed the allowed characters in metric names
 * added a clear deprecation policy
 * added unit tests for `Proxy`
+* refactored and improved `Proxy`