+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 
 * more pydoc for `satella.files`
 * add `try_unlink`
+* add `one_tuple`