+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # v2.9.6
 
 * `OmniHashableMixin` will be faster now
+* `OmniHashableMixin` will now allow single strings to be it's argument