+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 
 * `OmniHashableMixin` will be faster now
 * `OmniHashableMixin` will now allow single strings to be it's argument
+* added `SingleStartThread`