+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 * `OmniHashableMixin` will be faster now
 * `OmniHashableMixin` will now allow single strings to be it's argument
 * added `SingleStartThread`
+* `SelfCleaningDefaultDict` semantics slightly altered