+++ CHANGELOG.md
@@ -4,3 +4,4 @@
 * `OmniHashableMixin` will now allow single strings to be it's argument
 * added `SingleStartThread`
 * `SelfCleaningDefaultDict` semantics slightly altered
+* extra routines for `AtomicNumber`