+++ CHANGELOG.md
@@ -1,3 +1,5 @@
 # v2.9.7
 
+* add binary shift operations to `AtomicNumber`
+